@@ -93,14 +93,14 @@ class DefaultAlertDialogContent extends StatelessWidget {
   final Widget indicatorIcon;
   final Color indicatorBackgroundColor;
   final String title;
-  final String subtitle;
+  final String? subtitle;
   final List<Widget> buttons;
   const DefaultAlertDialogContent({
     Key? key,
     required this.indicatorIcon,
     this.indicatorBackgroundColor = Colors.transparent,
     required this.title,
-    required this.subtitle,
+    this.subtitle,
     required this.buttons,
   }) : super(key: key);
 
@@ -126,14 +126,15 @@ class DefaultAlertDialogContent extends StatelessWidget {
             style: TEXTSTYLES.kHeaderLarge.copyWith(color: COLORS.kDarkBlueColor),
           ),
           CustomSizes.smallSizedBoxHeight,
-          Text(
-            subtitle,
-            textAlign: TextAlign.center,
-            style: TEXTSTYLES.kLinkSmall.copyWith(
-              fontWeight: FONTS.Regular,
-              color: COLORS.kTextBlueColor,
+          if (subtitle != null)
+            Text(
+              subtitle!,
+              textAlign: TextAlign.center,
+              style: TEXTSTYLES.kLinkSmall.copyWith(
+                fontWeight: FONTS.Regular,
+                color: COLORS.kTextBlueColor,
+              ),
             ),
-          ),
           SizedBox(height: 24.w),
           ...buttons,
         ],
