@@ -4,8 +4,11 @@ import 'package:get/get.dart';
 import 'package:heyo/app/modules/chats/data/models/chat_model.dart';
 import 'package:heyo/app/modules/chats/widgets/chat_widget.dart';
 import 'package:heyo/app/modules/chats/widgets/empty_chats_widget.dart';
-import 'package:heyo/app/modules/shared/data/models/MessagesViewArgumentsModel.dart';
+import 'package:heyo/app/modules/shared/data/models/messages_view_arguments_model.dart';
+import 'package:heyo/app/modules/shared/utils/constants/colors.dart';
+import 'package:heyo/app/modules/shared/utils/constants/textStyles.dart';
 import 'package:heyo/app/routes/app_pages.dart';
+import 'package:heyo/generated/locales.g.dart';
 
 import '../controllers/chats_controller.dart';
 
@@ -15,8 +18,17 @@ class ChatsView extends GetView<ChatsController> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      backgroundColor: COLORS.kAppBackground,
+      appBar: AppBar(
+        backgroundColor: COLORS.kGreenMainColor,
+        title: Text(
+          LocaleKeys.HomePage_navbarItems_chats.tr,
+          style: TEXTSTYLES.kHeaderLarge,
+        ),
+        automaticallyImplyLeading: false,
+      ),
       body: Obx(
-        () => _buildChats(controller.chats.value),
+        () => _buildChats(controller.chats),
       ),
     );
   }
