@@ -12,8 +12,7 @@ import '../data/models/profile_model.dart';
 import '../data/models/user_model.dart';
 import '../widgets/invite_bttom_sheet.dart';
 
-class NewChatController extends GetxController
-    with GetSingleTickerProviderStateMixin {
+class NewChatController extends GetxController with GetSingleTickerProviderStateMixin {
   late AnimationController animController;
   late Animation<double> animation;
   late TextEditingController inputController;
@@ -41,8 +40,7 @@ class NewChatController extends GetxController
     inputController.addListener(() {
       searchUsers(inputController.text);
     });
-    nearbyUsers
-        .sort((a, b) => a.name.toLowerCase().compareTo(b.name.toLowerCase()));
+    nearbyUsers.sort((a, b) => a.name.toLowerCase().compareTo(b.name.toLowerCase()));
     super.onInit();
   }
 
@@ -74,24 +72,19 @@ class NewChatController extends GetxController
 
 // Mock data for user profile
   final profile = ProfileModel(
-      name: "Sample",
-      link: "https://heyo.core/m6ljkB4KJ",
-      walletAddress: "CB75...684A");
+      name: "Sample", link: "https://heyo.core/m6ljkB4KJ", walletAddress: "CB75...684A");
 
 // Mock data for the users
   final nearbyUsers = <UserModel>[
     UserModel(
       name: "Crapps Wallbanger",
       walletAddress: 'CB92...969A',
-      icon:
-          "https://raw.githubusercontent.com/Zunawe/identicons/HEAD/examples/poly.png",
-      Nickname: "Nickname",
+      icon: "https://raw.githubusercontent.com/Zunawe/identicons/HEAD/examples/poly.png",
+      nickname: "Nickname",
       chatModel: ChatModel(
         name: "Crapps Wallbanger",
-        icon:
-            "https://raw.githubusercontent.com/Zunawe/identicons/HEAD/examples/poly.png",
-        lastMessage:
-            "I'm still waiting for the reply. I'll let you know once they get back to me.",
+        icon: "https://raw.githubusercontent.com/Zunawe/identicons/HEAD/examples/poly.png",
+        lastMessage: "I'm still waiting for the reply. I'll let you know once they get back to me.",
         timestamp: "15:45",
       ),
     ),
@@ -104,8 +97,7 @@ class NewChatController extends GetxController
       chatModel: ChatModel(
         name: "Fancy Potato",
         icon: "https://avatars.githubusercontent.com/u/6634136?v=4",
-        lastMessage:
-            "I can arrange the meeting with her tomorrow if you're ok with that.",
+        lastMessage: "I can arrange the meeting with her tomorrow if you're ok with that.",
         timestamp: "Yesterday",
         isOnline: true,
         isVerified: true,
@@ -140,8 +132,7 @@ class NewChatController extends GetxController
     ),
   ].obs;
 
-  RefreshController refreshController =
-      RefreshController(initialRefresh: false);
+  RefreshController refreshController = RefreshController(initialRefresh: false);
   void onRefresh() async {
     await Future.delayed(Duration(milliseconds: 1000));
 
@@ -156,8 +147,7 @@ class NewChatController extends GetxController
       String inputedQuery = query.toLowerCase();
       return username.contains(inputedQuery);
     }).toList();
-    searchSuggestions
-        .sort((a, b) => a.name.toLowerCase().compareTo(b.name.toLowerCase()));
+    searchSuggestions.sort((a, b) => a.name.toLowerCase().compareTo(b.name.toLowerCase()));
     searchSuggestions.refresh();
   }
 
