@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:heyo/app/modules/new_chat/widgets/qr_scan_view.dart';
@@ -6,15 +5,18 @@ import 'package:heyo/app/modules/website-interact/website_interact_controller.da
 import 'package:heyo/generated/locales.g.dart';
 
 class WebsiteInteractView extends GetView<WebsiteInteractController> {
+  const WebsiteInteractView({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Material(
-        child: Container(
-          child: QrScanView(
-              hasBackButton: true,
-              subtitle: "",
-              title: LocaleKeys.Scan_Qr_title.tr,
-              onQRViewCreated: controller.handleScannedVal),
-        ));
+      child: Container(
+        child: QrScanView(
+            hasBackButton: true,
+            subtitle: "",
+            title: LocaleKeys.scanQrTitle.tr,
+            onQRViewCreated: controller.handleScannedVal),
+      ),
+    );
   }
 }
