@@ -10,11 +10,14 @@ class AppTranslation {
 
 class LocaleKeys {
   LocaleKeys._();
-  static const registration_WelcomePage_title = 'registration_WelcomePage_title';
-  static const registration_WelcomePage_subtitle = 'registration_WelcomePage_subtitle';
+  static const registration_WelcomePage_title =
+      'registration_WelcomePage_title';
+  static const registration_WelcomePage_subtitle =
+      'registration_WelcomePage_subtitle';
   static const registration_WelcomePage_buttons_Continue =
       'registration_WelcomePage_buttons_Continue';
-  static const registration_WelcomePage_buttons_Terms = 'registration_WelcomePage_buttons_Terms';
+  static const registration_WelcomePage_buttons_Terms =
+      'registration_WelcomePage_buttons_Terms';
   static const registration_BenefitsPage_sliders_title001 =
       'registration_BenefitsPage_sliders_title001';
   static const registration_BenefitsPage_sliders_subtitle001 =
@@ -31,32 +34,42 @@ class LocaleKeys {
       'registration_BenefitsPage_sliders_title004';
   static const registration_BenefitsPage_sliders_subtitle004 =
       'registration_BenefitsPage_sliders_subtitle004';
-  static const registration_BenefitsPage_divider = 'registration_BenefitsPage_divider';
-  static const registration_BenefitsPage_footer = 'registration_BenefitsPage_footer';
+  static const registration_BenefitsPage_divider =
+      'registration_BenefitsPage_divider';
+  static const registration_BenefitsPage_footer =
+      'registration_BenefitsPage_footer';
   static const registration_BenefitsPage_buttons_Create =
       'registration_BenefitsPage_buttons_Create';
   static const registration_BenefitsPage_buttons_SkipCreate =
       'registration_BenefitsPage_buttons_SkipCreate';
   static const registration_BenefitsPage_buttons_Whatis =
       'registration_BenefitsPage_buttons_Whatis';
-  static const registration_PINcodePage_title = 'registration_PINcodePage_title';
-  static const registration_PINcodePage_subtitle = 'registration_PINcodePage_subtitle';
-  static const registration_PINcodePage_pinType = 'registration_PINcodePage_pinType';
-  static const registration_PINcodePage_pinRepeat = 'registration_PINcodePage_pinRepeat';
-  static const registration_PINcodePage_dialogTitle = 'registration_PINcodePage_dialogTitle';
+  static const registration_PINcodePage_title =
+      'registration_PINcodePage_title';
+  static const registration_PINcodePage_subtitle =
+      'registration_PINcodePage_subtitle';
+  static const registration_PINcodePage_pinType =
+      'registration_PINcodePage_pinType';
+  static const registration_PINcodePage_pinRepeat =
+      'registration_PINcodePage_pinRepeat';
+  static const registration_PINcodePage_dialogTitle =
+      'registration_PINcodePage_dialogTitle';
   static const registration_PINcodePage_buttons_AddBiometric =
       'registration_PINcodePage_buttons_AddBiometric';
   static const registration_PINcodePage_buttons_SkipBiometric =
       'registration_PINcodePage_buttons_SkipBiometric';
-  static const registration_PrivateKeysPage_Generating = 'registration_PrivateKeysPage_Generating';
-  static const registration_PersonalizationPage_title = 'registration_PersonalizationPage_title';
+  static const registration_PrivateKeysPage_Generating =
+      'registration_PrivateKeysPage_Generating';
+  static const registration_PersonalizationPage_title =
+      'registration_PersonalizationPage_title';
   static const registration_PersonalizationPage_subtitle =
       'registration_PersonalizationPage_subtitle';
   static const registration_PersonalizationPage_buttons_Generate =
       'registration_PersonalizationPage_buttons_Generate';
   static const registration_PersonalizationPage_buttons_Continue =
       'registration_PersonalizationPage_buttons_Continue';
-  static const registration_RecoveryMethodPage_title = 'registration_RecoveryMethodPage_title';
+  static const registration_RecoveryMethodPage_title =
+      'registration_RecoveryMethodPage_title';
   static const registration_RecoveryMethodPage_firstOption_title =
       'registration_RecoveryMethodPage_firstOption_title';
   static const registration_RecoveryMethodPage_firstOption_subtitle =
@@ -75,53 +88,109 @@ class LocaleKeys {
   static const HomePage_navbarItems_wallet = 'HomePage_navbarItems_wallet';
   static const HomePage_navbarItems_search = 'HomePage_navbarItems_search';
   static const HomePage_navbarItems_account = 'HomePage_navbarItems_account';
-  static const HomePage_Chats_emptyState_title = 'HomePage_Chats_emptyState_title';
-  static const HomePage_Chats_emptyState_subtitle = 'HomePage_Chats_emptyState_subtitle';
-  static const HomePage_Chats_emptyState_invite = 'HomePage_Chats_emptyState_invite';
+  static const HomePage_Chats_emptyState_title =
+      'HomePage_Chats_emptyState_title';
+  static const HomePage_Chats_emptyState_subtitle =
+      'HomePage_Chats_emptyState_subtitle';
+  static const HomePage_Chats_emptyState_invite =
+      'HomePage_Chats_emptyState_invite';
+  static const HomePage_Calls_emptyState_title =
+      'HomePage_Calls_emptyState_title';
+  static const HomePage_Calls_emptyState_subtitle =
+      'HomePage_Calls_emptyState_subtitle';
+  static const HomePage_Calls_hour = 'HomePage_Calls_hour';
+  static const HomePage_Calls_hours = 'HomePage_Calls_hours';
+  static const HomePage_Calls_minute = 'HomePage_Calls_minute';
+  static const HomePage_Calls_minutes = 'HomePage_Calls_minutes';
+  static const HomePage_Calls_durationAgo = 'HomePage_Calls_durationAgo';
+  static const HomePage_Calls_deleteCallDialog_title =
+      'HomePage_Calls_deleteCallDialog_title';
+  static const HomePage_Calls_deleteCallDialog_delete =
+      'HomePage_Calls_deleteCallDialog_delete';
+  static const HomePage_Calls_deleteCallDialog_cancel =
+      'HomePage_Calls_deleteCallDialog_cancel';
+  static const HomePage_Calls_deleteAllCallsDialog_title =
+      'HomePage_Calls_deleteAllCallsDialog_title';
+  static const HomePage_Calls_deleteAllCallsDialog_delete =
+      'HomePage_Calls_deleteAllCallsDialog_delete';
+  static const HomePage_Calls_deleteAllCallsDialog_cancel =
+      'HomePage_Calls_deleteAllCallsDialog_cancel';
+  static const HomePage_Calls_bottomSheet_deleteAllCalls =
+      'HomePage_Calls_bottomSheet_deleteAllCalls';
   static const HomePage_bottomSheet_newChat = 'HomePage_bottomSheet_newChat';
   static const HomePage_bottomSheet_newGroup = 'HomePage_bottomSheet_newGroup';
   static const HomePage_bottomSheet_invite = 'HomePage_bottomSheet_invite';
+  static const CallHistory_appbar = 'CallHistory_appbar';
+  static const CallHistory_missedCall = 'CallHistory_missedCall';
+  static const CallHistory_outgoing = 'CallHistory_outgoing';
+  static const CallHistory_incoming = 'CallHistory_incoming';
+  static const CallHistory_callDeclined = 'CallHistory_callDeclined';
+  static const CallHistory_callCanceled = 'CallHistory_callCanceled';
+  static const CallHistory_notAnswered = 'CallHistory_notAnswered';
+  static const CallPage_participants = 'CallPage_participants';
+  static const CallPage_ringing = 'CallPage_ringing';
+  static const CallPage_currentlyOffline = 'CallPage_currentlyOffline';
+  static const CallPage_sendAMessageButton = 'CallPage_sendAMessageButton';
+  static const CallPage_close = 'CallPage_close';
   static const newChat_newChatAppBar = 'newChat_newChatAppBar';
   static const newChat_slider_contacts = 'newChat_slider_contacts';
   static const newChat_slider_nearbyUsers = 'newChat_slider_nearbyUsers';
   static const newChat_usernameInputFiled = 'newChat_usernameInputFiled';
-  static const newChat_emptyStateTitleContacts = 'newChat_emptyStateTitleContacts';
-  static const newChat_emptyStateTitleNearbyUsers = 'newChat_emptyStateTitleNearbyUsers';
+  static const newChat_emptyStateTitleContacts =
+      'newChat_emptyStateTitleContacts';
+  static const newChat_emptyStateTitleNearbyUsers =
+      'newChat_emptyStateTitleNearbyUsers';
   static const newChat_usernameInput = 'newChat_usernameInput';
   static const newChat_searchResults = 'newChat_searchResults';
   static const newChat_buttons_invite = 'newChat_buttons_invite';
   static const newChat_buttons_filter = 'newChat_buttons_filter';
-  static const newChat_buttons_pullDownToRefresh = 'newChat_buttons_pullDownToRefresh';
+  static const newChat_buttons_pullDownToRefresh =
+      'newChat_buttons_pullDownToRefresh';
   static const newChat_userBottomSheet_chat = 'newChat_userBottomSheet_chat';
-  static const newChat_userBottomSheet_addToContacts = 'newChat_userBottomSheet_addToContacts';
+  static const newChat_userBottomSheet_addToContacts =
+      'newChat_userBottomSheet_addToContacts';
   static const newChat_userBottomSheet_block = 'newChat_userBottomSheet_block';
-  static const newChat_userBottomSheet_userInfo = 'newChat_userBottomSheet_userInfo';
+  static const newChat_userBottomSheet_userInfo =
+      'newChat_userBottomSheet_userInfo';
   static const newChat_inviteBottomSheet_inviteYourFriend =
       'newChat_inviteBottomSheet_inviteYourFriend';
-  static const newChat_inviteBottomSheet_shareLink = 'newChat_inviteBottomSheet_shareLink';
+  static const newChat_inviteBottomSheet_shareLink =
+      'newChat_inviteBottomSheet_shareLink';
   static const MessagesPage_textFieldHint = 'MessagesPage_textFieldHint';
-  static const MessagesPage_endToEndEncryptedMessaging = 'MessagesPage_endToEndEncryptedMessaging';
+  static const MessagesPage_endToEndEncryptedMessaging =
+      'MessagesPage_endToEndEncryptedMessaging';
   static const MessagesPage_onlineVia = 'MessagesPage_onlineVia';
-  static const MessagesPage_countSelectedMessages = 'MessagesPage_countSelectedMessages';
+  static const MessagesPage_countSelectedMessages =
+      'MessagesPage_countSelectedMessages';
   static const MessagesPage_replyingTo = 'MessagesPage_replyingTo';
   static const MessagesPage_replyToImage = 'MessagesPage_replyToImage';
   static const MessagesPage_replyToVideo = 'MessagesPage_replyToVideo';
   static const MessagesPage_replyToAudio = 'MessagesPage_replyToAudio';
   static const MessagesPage_forwarded = 'MessagesPage_forwarded';
   static const MessagesPage_sharedLocation = 'MessagesPage_sharedLocation';
-  static const MessagesPage_sharingLiveLocation = 'MessagesPage_sharingLiveLocation';
-  static const MessagesPage_liveShareRemainingMinute = 'MessagesPage_liveShareRemainingMinute';
-  static const MessagesPage_liveShareRemainingMinutes = 'MessagesPage_liveShareRemainingMinutes';
-  static const MessagesPage_liveShareRemainingHour = 'MessagesPage_liveShareRemainingHour';
-  static const MessagesPage_liveShareRemainingHours = 'MessagesPage_liveShareRemainingHours';
-  static const MessagesPage_liveShareLessThanMinute = 'MessagesPage_liveShareLessThanMinute';
-  static const MessagesPage_liveShareRemainingTime = 'MessagesPage_liveShareRemainingTime';
-  static const MessagesPage_finishedLiveShareTitle = 'MessagesPage_finishedLiveShareTitle';
+  static const MessagesPage_sharingLiveLocation =
+      'MessagesPage_sharingLiveLocation';
+  static const MessagesPage_liveShareRemainingMinute =
+      'MessagesPage_liveShareRemainingMinute';
+  static const MessagesPage_liveShareRemainingMinutes =
+      'MessagesPage_liveShareRemainingMinutes';
+  static const MessagesPage_liveShareRemainingHour =
+      'MessagesPage_liveShareRemainingHour';
+  static const MessagesPage_liveShareRemainingHours =
+      'MessagesPage_liveShareRemainingHours';
+  static const MessagesPage_liveShareLessThanMinute =
+      'MessagesPage_liveShareLessThanMinute';
+  static const MessagesPage_liveShareRemainingTime =
+      'MessagesPage_liveShareRemainingTime';
+  static const MessagesPage_finishedLiveShareTitle =
+      'MessagesPage_finishedLiveShareTitle';
   static const MessagesPage_finishedLiveShareOtherTitle =
       'MessagesPage_finishedLiveShareOtherTitle';
-  static const MessagesPage_finishedLiveShareSubtitle = 'MessagesPage_finishedLiveShareSubtitle';
+  static const MessagesPage_finishedLiveShareSubtitle =
+      'MessagesPage_finishedLiveShareSubtitle';
   static const MessagesPage_stopSharing = 'MessagesPage_stopSharing';
-  static const MessagesPage_deleteMessagesDialog_title = 'MessagesPage_deleteMessagesDialog_title';
+  static const MessagesPage_deleteMessagesDialog_title =
+      'MessagesPage_deleteMessagesDialog_title';
   static const MessagesPage_deleteMessagesDialog_titlePlural =
       'MessagesPage_deleteMessagesDialog_titlePlural';
   static const MessagesPage_deleteMessagesDialog_subtitle =
@@ -138,15 +207,20 @@ class LocaleKeys {
       'MessagesPage_deleteMessagesDialog_deleteForMe';
   static const forwardMassagesPage_appbar = 'forwardMassagesPage_appbar';
   static const forwardMassagesPage_textInput = 'forwardMassagesPage_textInput';
-  static const forwardMassagesPage_recentContacts = 'forwardMassagesPage_recentContacts';
-  static const forwardMassagesPage_OtherContacts = 'forwardMassagesPage_OtherContacts';
+  static const forwardMassagesPage_recentContacts =
+      'forwardMassagesPage_recentContacts';
+  static const forwardMassagesPage_OtherContacts =
+      'forwardMassagesPage_OtherContacts';
   static const forwardMassagesPage_contacts = 'forwardMassagesPage_contacts';
-  static const forwardMassagesPage_bottomBar_forwardTO = 'forwardMassagesPage_bottomBar_forwardTO';
-  static const forwardMassagesPage_bottomBar_messages = 'forwardMassagesPage_bottomBar_messages';
+  static const forwardMassagesPage_bottomBar_forwardTO =
+      'forwardMassagesPage_bottomBar_forwardTO';
+  static const forwardMassagesPage_bottomBar_messages =
+      'forwardMassagesPage_bottomBar_messages';
   static const permissionDialog_continue = 'permissionDialog_continue';
   static const permissionDialog_notNow = 'permissionDialog_notNow';
   static const permissionDialog_allowAccess = 'permissionDialog_allowAccess';
-  static const permissionDialog_locationSubtitle = 'permissionDialog_locationSubtitle';
+  static const permissionDialog_locationSubtitle =
+      'permissionDialog_locationSubtitle';
   static const shareLocationPage_appBar = 'shareLocationPage_appBar';
   static const shareLocationPage_bottomSheet_shareLiveLocation =
       'shareLocationPage_bottomSheet_shareLiveLocation';
@@ -169,11 +243,13 @@ class LocaleKeys {
   static const forward = 'forward';
   static const delete = 'delete';
   static const offline = 'offline';
+  static const scanQrTitle = 'scanQrTitle';
   static const AddContacts_addToContacts = 'AddContacts_addToContacts';
   static const AddContacts_addNickname = 'AddContacts_addNickname';
-  static const AddContacts_AddNicknameSubtitle = 'AddContacts_AddNicknameSubtitle';
-  static const AddContacts_buttons_addToContacts = 'AddContacts_buttons_addToContacts';
-  static const Scan_Qr_title = 'scan_qr_title';
+  static const AddContacts_AddNicknameSubtitle =
+      'AddContacts_AddNicknameSubtitle';
+  static const AddContacts_buttons_addToContacts =
+      'AddContacts_buttons_addToContacts';
 }
 
 class Locales {
@@ -183,16 +259,19 @@ class Locales {
         'A truly private messaging app built on a decentralized technology, secured by CorePass.',
     'registration_WelcomePage_buttons_Continue': 'Continue with CoreID',
     'registration_WelcomePage_buttons_Terms': 'Terms & Privacy Policy',
-    'registration_BenefitsPage_sliders_title001': 'One account \n for all your services',
+    'registration_BenefitsPage_sliders_title001':
+        'One account \n for all your services',
     'registration_BenefitsPage_sliders_subtitle001':
         'Keep your valuable data safely stored \n and quickly accesible from one place.',
-    'registration_BenefitsPage_sliders_title002': 'Assurance that you\'ve \n got the right person',
+    'registration_BenefitsPage_sliders_title002':
+        'Assurance that you\'ve \n got the right person',
     'registration_BenefitsPage_sliders_subtitle002':
         'Know the person you are connecting with \n and how your data is being used.',
     'registration_BenefitsPage_sliders_title003': 'Your data has value',
     'registration_BenefitsPage_sliders_subtitle003':
         'Each time your data gets requested, you \n earn an income. Your identity becomes \n your saving and spending tool.',
-    'registration_BenefitsPage_sliders_title004': 'Keep your earnings \n in a safe place',
+    'registration_BenefitsPage_sliders_title004':
+        'Keep your earnings \n in a safe place',
     'registration_BenefitsPage_sliders_subtitle004':
         'CorePass comes with an integrated \n cryptocurrency wallet to keep your \n savings safe and secure.',
     'registration_BenefitsPage_divider': 'OR',
@@ -201,23 +280,29 @@ class Locales {
     'registration_BenefitsPage_buttons_SkipCreate': 'I already have CoreID',
     'registration_BenefitsPage_buttons_Whatis': 'What is CoreID?',
     'registration_PINcodePage_title': 'Set PIN',
-    'registration_PINcodePage_subtitle': 'Set a 6 digit PIN code to protect your Heyo application',
+    'registration_PINcodePage_subtitle':
+        'Set a 6 digit PIN code to protect your Heyo application',
     'registration_PINcodePage_pinType': 'Type your PIN code',
     'registration_PINcodePage_pinRepeat': 'Repeat your PIN code',
-    'registration_PINcodePage_dialogTitle': 'Add an extra layer to your security',
+    'registration_PINcodePage_dialogTitle':
+        'Add an extra layer to your security',
     'registration_PINcodePage_buttons_AddBiometric': 'Add biometric log in',
-    'registration_PINcodePage_buttons_SkipBiometric': 'Continue without this option',
+    'registration_PINcodePage_buttons_SkipBiometric':
+        'Continue without this option',
     'registration_PrivateKeysPage_Generating': 'Generating private keys...',
     'registration_PersonalizationPage_title': 'Choose your chat name',
     'registration_PersonalizationPage_subtitle':
         'All users will see only your CoreID and a readable name that can’t be changed.',
     'registration_PersonalizationPage_buttons_Generate': 'Generate new names',
     'registration_PersonalizationPage_buttons_Continue': 'Continue',
-    'registration_RecoveryMethodPage_title': 'How do you want to back up your account?',
-    'registration_RecoveryMethodPage_firstOption_title': 'Offline, using recovery phrase',
+    'registration_RecoveryMethodPage_title':
+        'How do you want to back up your account?',
+    'registration_RecoveryMethodPage_firstOption_title':
+        'Offline, using recovery phrase',
     'registration_RecoveryMethodPage_firstOption_subtitle':
         'A 24 word phrase will be generated. We highly recommend you to write these words down on paper and store them in a safe offline place.',
-    'registration_RecoveryMethodPage_SecondOption_title': 'Digitally, using private key',
+    'registration_RecoveryMethodPage_SecondOption_title':
+        'Digitally, using private key',
     'registration_RecoveryMethodPage_SecondOption_subtitle':
         'Your CoreID account can be recovered with a private key which you will have to protect with a password and store.',
     'registration_RecoveryMethodPage_buttons_Skip': 'Skip for now',
@@ -229,11 +314,40 @@ class Locales {
     'HomePage_navbarItems_search': 'Search',
     'HomePage_navbarItems_account': 'Account',
     'HomePage_Chats_emptyState_title': 'It looks empty here',
-    'HomePage_Chats_emptyState_subtitle': 'Start with inviting your friends or start new chats',
+    'HomePage_Chats_emptyState_subtitle':
+        'Start with inviting your friends or start new chats',
     'HomePage_Chats_emptyState_invite': 'Invite',
+    'HomePage_Calls_emptyState_title': 'You have no calls',
+    'HomePage_Calls_emptyState_subtitle':
+        'Call someone from your contact list by hitting that + button',
+    'HomePage_Calls_hour': '@count hour',
+    'HomePage_Calls_hours': '@count hours',
+    'HomePage_Calls_minute': '@count minute',
+    'HomePage_Calls_minutes': '@count minutes',
+    'HomePage_Calls_durationAgo': '@duration ago',
+    'HomePage_Calls_deleteCallDialog_title': 'Delete this call from history?',
+    'HomePage_Calls_deleteCallDialog_delete': 'Delete',
+    'HomePage_Calls_deleteCallDialog_cancel': 'Cancel',
+    'HomePage_Calls_deleteAllCallsDialog_title':
+        'Delete all calls from history?',
+    'HomePage_Calls_deleteAllCallsDialog_delete': 'Delete',
+    'HomePage_Calls_deleteAllCallsDialog_cancel': 'Cancel',
+    'HomePage_Calls_bottomSheet_deleteAllCalls': 'Delete all calls',
     'HomePage_bottomSheet_newChat': 'Start new chat',
     'HomePage_bottomSheet_newGroup': 'Start new group',
     'HomePage_bottomSheet_invite': 'Invite',
+    'CallHistory_appbar': 'Call history',
+    'CallHistory_missedCall': 'Missed call',
+    'CallHistory_outgoing': 'Outgoing',
+    'CallHistory_incoming': 'Incoming',
+    'CallHistory_callDeclined': 'Call declined',
+    'CallHistory_callCanceled': 'Call canceled',
+    'CallHistory_notAnswered': 'Not answered',
+    'CallPage_participants': 'Participants',
+    'CallPage_ringing': 'Ringing...',
+    'CallPage_currentlyOffline': 'Currently offline',
+    'CallPage_sendAMessageButton': 'Send a message instead',
+    'CallPage_close': 'Close',
     'newChat_newChatAppBar': 'New Chat',
     'newChat_slider_contacts': 'Contacts',
     'newChat_slider_nearbyUsers': 'Nearby users',
@@ -277,10 +391,12 @@ class Locales {
     'MessagesPage_deleteMessagesDialog_title': 'Delete message?',
     'MessagesPage_deleteMessagesDialog_titlePlural': 'Delete @count messages?',
     'MessagesPage_deleteMessagesDialog_subtitle': 'Delete this message?',
-    'MessagesPage_deleteMessagesDialog_subtitlePlural': 'Delete these messages?',
+    'MessagesPage_deleteMessagesDialog_subtitlePlural':
+        'Delete these messages?',
     'MessagesPage_deleteMessagesDialog_cancel': 'Cancel',
     'MessagesPage_deleteMessagesDialog_delete': 'Delete',
-    'MessagesPage_deleteMessagesDialog_deleteForEveryone': 'Delete for everyone',
+    'MessagesPage_deleteMessagesDialog_deleteForEveryone':
+        'Delete for everyone',
     'MessagesPage_deleteMessagesDialog_deleteForMe': 'Delete only for me',
     'forwardMassagesPage_appbar': 'Forward to',
     'forwardMassagesPage_textInput': 'Search in contacts',
@@ -299,7 +415,8 @@ class Locales {
     'shareLocationPage_bottomSheet_sendPickerLocation': 'Send Picker Location',
     'shareLocationPage_bottomSheet_locationAccessDisclaimer':
         'We will access your location only for the duration you chosen and you can stop sharing it any time. We won’t use your location for any purpose other than sharing it in this chat.',
-    'shareLocationPage_bottomSheet_shareLiveLocationWith': 'Share my live location with @name for:',
+    'shareLocationPage_bottomSheet_shareLiveLocationWith':
+        'Share my live location with @name for:',
     'shareLocationPage_bottomSheet_shareDurationMinutes': '@count mins',
     'shareLocationPage_bottomSheet_shareDurationHour': '@count hour',
     'shareLocationPage_bottomSheet_shareDurationHours': '@count hours',
@@ -310,6 +427,7 @@ class Locales {
     'forward': 'Forward',
     'delete': 'Delete',
     'offline': 'Offline',
+    'scanQrTitle': 'Scan Qr Code',
     'AddContacts_addToContacts': 'Add to contacts',
     'AddContacts_addNickname': 'Add nickname',
     'AddContacts_AddNicknameSubtitle':
